Split the local variable check into two parts rather than 1. 1. If there is at least one local variable read within the return statement. 2. If there is at least one local variable written within the for loop Note: These variables now do NOT have to be the same - simplicity in the check.
Also included a check to ensure at least one local varaible is defined
within the method. This refers to an immediate local variable (i.e. at
the beginning of the method).

TODO: If later revisions require a check for any local variable declared
and used. But the pattern tends to aim at simplicity therefore the
current implementation suffices.